requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
python-multipart
pdf2image
Pillow
httpx
google-generativeai
gigachat
tenacity
jinja2
python-jose[cryptography]
passlib[bcrypt]
pytesseract
requests
fpdf
